Allow specifying KVO options for `NSObject.values(forKeyPath:)`.
Original motivation for this is you don't always want the initial value.

A real-world example is from dealing with an `AVAsset` in `AVFoundation`.
When you read `asset.duration` and the duration is not known yet, then
`AVAsset` will busy-load asset metadata to give you a value, which can
lock up the main thread. You can avoid this by reading the status of the
`duration` key using `AVAsynchronousKeyValueLoading`, and then use KVO to
deliver updates if the duration is not yet known.